// src/laana/types/event/wrapper.ts
export interface EventWrapper {
    /**
     * @generated from protobuf field: string eventId = 1;
     */
    eventId: string;
    /**
     * @generated from protobuf field: uint64 time = 2;
     */
    time: bigint;
    /**
     * @generated from protobuf field: string botUin = 3;
     */
    botUin: string;
    /**
     * @generated from protobuf oneof: event
     */
    event: {
        oneofKind: "buddyAddedEvent";
        /**
         * @generated from protobuf field: Laana.BuddyAddedEvent buddyAddedEvent = 101;
         */
        buddyAddedEvent: BuddyAddedEvent;
    } | {
        oneofKind: "buddyMessageWithdrawEvent";
        /**
         * @generated from protobuf field: Laana.BuddyMessageWithdrawEvent buddyMessageWithdrawEvent = 102;
         */
        buddyMessageWithdrawEvent: BuddyMessageWithdrawEvent;
    } | {
        oneofKind: "groupAdminChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupAdminChangeEvent groupAdminChangeEvent = 201;
         */
        groupAdminChangeEvent: GroupAdminChangeEvent;
    } | {
        oneofKind: "groupEssenceMessageEvent";
        /**
         * @generated from protobuf field: Laana.GroupEssenceMessageEvent groupEssenceMessageEvent = 202;
         */
        groupEssenceMessageEvent: GroupEssenceMessageEvent;
    } | {
        oneofKind: "groupMemberCardChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberCardChangeEvent groupMemberCardChangeEvent = 203;
         */
        groupMemberCardChangeEvent: GroupMemberCardChangeEvent;
    } | {
        oneofKind: "groupMemberIncreaseEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberIncreaseEvent groupMemberIncreaseEvent = 204;
         */
        groupMemberIncreaseEvent: GroupMemberIncreaseEvent;
    } | {
        oneofKind: "groupMemberDecreaseEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberDecreaseEvent groupMemberDecreaseEvent = 205;
         */
        groupMemberDecreaseEvent: GroupMemberDecreaseEvent;
    } | {
        oneofKind: "groupMemberTitleChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberTitleChangeEvent groupMemberTitleChangeEvent = 206;
         */
        groupMemberTitleChangeEvent: GroupMemberTitleChangeEvent;
    } | {
        oneofKind: "groupMessageWithdrawEvent";
        /**
         * @generated from protobuf field: Laana.GroupMessageWithdrawEvent groupMessageWithdrawEvent = 207;
         */
        groupMessageWithdrawEvent: GroupMessageWithdrawEvent;
    } | {
        oneofKind: "groupMuteMemberEvent";
        /**
         * @generated from protobuf field: Laana.GroupShutUpMemberEvent groupMuteMemberEvent = 208;
         */
        groupMuteMemberEvent: GroupShutUpMemberEvent;
    } | {
        oneofKind: "groupMuteAllEvent";
        /**
         * @generated from protobuf field: Laana.GroupShutUpAllEvent groupMuteAllEvent = 209;
         */
        groupMuteAllEvent: GroupShutUpAllEvent;
    } | {
        oneofKind: "groupUnmuteMemberEvent";
        /**
         * @generated from protobuf field: Laana.GroupLiftShutUpMemberEvent groupUnmuteMemberEvent = 210;
         */
        groupUnmuteMemberEvent: GroupLiftShutUpMemberEvent;
    } | {
        oneofKind: "groupUnmuteAllEvent";
        /**
         * @generated from protobuf field: Laana.GroupLiftShutUpAllEvent groupUnmuteAllEvent = 211;
         */
        groupUnmuteAllEvent: GroupLiftShutUpAllEvent;
    } | {
        oneofKind: "groupEmojiLikeEvent";
        /**
         * @generated from protobuf field: Laana.GroupEmojiLikeEvent groupEmojiLikeEvent = 212;
         */
        groupEmojiLikeEvent: GroupEmojiLikeEvent;
    } | {
        oneofKind: "someoneRequestToAddBuddyEvent";
        /**
         * @generated from protobuf field: Laana.SomeoneRequestToAddBuddyEvent someoneRequestToAddBuddyEvent = 301;
         */
        someoneRequestToAddBuddyEvent: SomeoneRequestToAddBuddyEvent;
    } | {
        oneofKind: "someoneRequestToJoinGroupEvent";
        /**
         * @generated from protobuf field: Laana.SomeoneRequestToJoinGroupEvent someoneRequestToJoinGroupEvent = 302;
         */
        someoneRequestToJoinGroupEvent: SomeoneRequestToJoinGroupEvent;
    } | {
        oneofKind: "selfInvitedToGroupEvent";
        /**
         * @generated from protobuf field: Laana.SelfInvitedToGroupEvent selfInvitedToGroupEvent = 303;
         */
        selfInvitedToGroupEvent: SelfInvitedToGroupEvent;
    } | {
        oneofKind: "extendedEvent";
        /**
         * @generated from protobuf field: Laana.ExtendedEvent extendedEvent = 9999;
         */
        extendedEvent: ExtendedEvent;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message Laana.ExtendedEvent
 */
export interface ExtendedEvent {
    /**
     * @generated from protobuf field: string eventName = 1;
     */
    eventName: string;
    /**
     * @generated from protobuf field: bytes eventData = 2;
     */
    eventData: Uint8Array;
}
// ...
export const EventWrapper = new EventWrapper$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ExtendedEvent$Type extends MessageType<ExtendedEvent> {
    constructor() {
        super("Laana.ExtendedEvent", [
            { no: 1, name: "eventName", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "eventData", kind: "scalar", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
    create(value?: PartialMessage<ExtendedEvent>): ExtendedEvent {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.eventName = "";
        message.eventData = new Uint8Array(0);
        if (value !== undefined)
            reflectionMergePartial<ExtendedEvent>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ExtendedEvent): ExtendedEvent {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string eventName */ 1:
                    message.eventName = reader.string();
                    break;
                case /* bytes eventData */ 2:
                    message.eventData = reader.bytes();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ExtendedEvent, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string eventName = 1; */
        if (message.eventName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.eventName);
        /* bytes eventData = 2; */
        if (message.eventData.length)
            writer.tag(2, WireType.LengthDelimited).bytes(message.eventData);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Laana.ExtendedEvent
 */
export const ExtendedEvent = new ExtendedEvent$Type();